feat: Export getColorName and getColorIndex
